CSS Updates for CustomerEdit page and customer details page
I changed the CSS on the customer details page to make it look more formal and centered, because before it was too small and the table was too long. Also, I created an external CSS file for the customerEdit page so that the code can follow better coding standards. And I changed the CSS to make the form more larger because it was very small before.